Align IngenPlanerCard for SM and AG
